refactor: Refatora classe produto
